Separate Data from process/functions
Previously, the only files were Class.js and Data.js.

Data.js was originally meant to contain only the data of each new Problem class. However, the search logic was created on the Data.js file as a proof of concept, before creating a separate .js file, and having to deal with connecting modules.

Since the search loop concept was proven, this commit has moved the search function to App.js.