Updated Template, added Problems, corrected method console.log
Updated the template in Class.js, and updated comments on existing problems in Data.js to add notes regarding where to add test code, my code, and best-practice code.

Added more previously completed Problems.

In App.js
Added count variable, so that logged match # is an iteration as matches are found. (previously, it was based on for loop i, which was actually printing out the i iteration number (+1)

Revised variable name used in for loop, previously 'methods'. Changed to 'met'. This is because methods is a key value in my Problems Class, and that key is directly used within the same loop logic. There were no errors previously, but the change to 'met' is to avoid confusion.

In the c loop,
Revised console.log of methods:
1) Revised to 'met' since met = the same reference previously made.
2) Added .join(", ") to fix formatting issue when the methods list is logged. (now has space after each comma).